requirements_nv.txt
torch --extra-index-url https://download.pytorch.org/whl/cu124
torchvision
torchaudio
llama-index
llama-index-llms-ollama
ipython
llama-index-graph-stores-neo4j
llama-index-vector-stores-neo4jvector
networkx
pyvis
gradio
neo4j
ipykernel
docker
transformers
huggingface_hub[inference]
sentence-transformers
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-llms-huggingface-api
optimum
bitsandbytes
protobuf
sentencepiece
mistralai
mauve-text
llama-index-llms-mistralai
llama-index-embeddings-mistralai
pandas
openpyxl
pdfplumber
sentence-transformers
accelerate
transformers 
llama-index-llms-huggingface
bitsandbytes